Stop aria2c on download completion
With commit f3c352a, we are now enabling aria2c's RPC interface which allows us to
query it to get download statuses (using aria2p cli tool for instance).
This RPC option (although it's not mentioned in the doc) changes the behavior of aria2c
and turns it into a daemon, even if it was launched with a single download.
It thus doesn't stop anymore at the end of said download.

This introduce two changes:
- a script that terminates all aria2c processes. It uses SIGHUP as aria2c handles it
as a request to terminate, ensuring exit code corresponds to aria2c's status.
- using said script as download-complete and download-error hooks